removed old modal containers, compoennts and other unecessary files
