+++ src/js/components/Loader/index.js
