+++ src/js/pages/root.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import setupSubscriptions from 'redux-subscriptions';
-import configureStore from '../utils/configureStore';
+import configureStore from '../redux/configureStore';
 import Application from './app';
-import { loadState, saveState } from '../utils/localStorage';
+import { loadState, saveState } from '../localStorage';
 
 //loading persistedState from filesystem using loadState()
 const persistedState = loadState();
